fixed Plant.id API generation

+++ api/generate_sdk.sh
@@ -1,6 +1,12 @@
 #!/usr/bin/env sh
 # add  "templateDir": "sdk/templates/py" to JSON or -t sdk/templates/py to CLI
 
+if [[ ! -d node_modules ]]; then
+	echo Run npm install first
+	exit 1
+fi
+
 rm -rf sdk/py
-npx p2o Plant.id_v3.postman_collection.json -o sdk/p2o_options.json >sdk/Plant.id_v3_API.spec.yml 
-npx openapi-generator-cli generate -g python -i sdk/Plant.id_v3_API.spec.yml -o sdk/py -c sdk/py_conf.json
+tempyml=$(mktemp)
+npx p2o Plant.id_v3.postman_collection.json -o p2o_options.json >"${tempyml}"
+npx openapi-generator-cli generate -g python -i "${tempyml}" -o sdk/py -c py_conf.json
